Debug edge case of lowestBitClear
For a fixed length or byte string type, all bits set mean lowestBitClear
should return the length in bits, not -1.
But the "obvious" formula using findFirstSetBit returns -1 because
we were complementing with all bits, not with a notional infinite  stream
of bits as should be the ideal case.
Inserting special case to recognize that.